Several corrections detailed on the description
- Printing initialization messages.
- Asking information to the user at the beginning and storing it in
variables, so the user does not need to edit the script.
- Verifying if Bash is being used, because the script only works with
Bash, and changing to Bash when not in use.
- Sending some text to some commands, so the user does not need to deal
with them.
- Uncommenting the lines in 'vars' that must be updated by the user.
- Assignment of "$BEFORE_RULES" corrected to assignment of
"BEFORE_RULES".
- Using a directory accessible by 'root' and using 'root' to perform the
installation (this solves some problems with authorization access to
files).
- Removed some unnecessary lines that were on the tutorial because the
DigitalOcean's tutorial has the CA on one machine and the OpenVPN server
on another and the script is configured to do all in the same machine.
- 2 TODOs added.