Code moved to Python
Transition to Python started on July 19, 2018, 15h/16h and the major part has been done on the same day. Only adjustments and fixes have been done the rest of the time.
Motivation:
On July 17, 2018, I found that the variables "is_valid_port", "is_valid_ipv4_address", "is_valid_ipv6_address" and "is_valid_dns_name" were not being treated as variables in the ifs used on the verifications, because I forgot to add "$" to the beginning of them. After fixing this, I found that other things were not working as expected on the functions that created the content of this variables, so I decided to move to a fully-featured language and decided to give a try to Python that I've never used and appeared to be a good option.

Changed:
- Shell scripts moved to Python, except "make_config.sh".
- Misspellings of "instalation" corrected to "installation".
- Code divided into files.
- Python code is verifying if the executed code and subprocesses execute as expected so the user does not need to check the output (with this, the message "Pay attention to the output for errors or warnings." has been removed).
- "VARS_DIR" renamed to "vars_path".
- "INITIAL_PWD" renamed to "initial_cwd".
- Not copying "configuration_variables.py" to the installation directory and not calling "nano" to update the content of the file (with this, the message related to updating the file has been updated).
- README.md updated.

Limitations:
- When elevating privileges calling "sudo" assumes that "python3" is installed.
- Clients passwords are not being sent to the program that generates the client certificate and key pair, therefore, the user needs to indicate them again on the shell.